Support health check headers (#598)
* bump go-fastly to v6.8.0

* support changes in go-fastly bumped version

* resolve conflict 1

* fix unit tests to support changes in go-fastly bump

* update docs

* NOTE about sorting for sake of test validation

* fix linter issues

* update docs

* add alpha message

+++ fastly/block_fastly_service_logging_blobstorage_test.go
@@ -51,7 +51,7 @@ func TestResourceFastlyFlattenBlobStorage(t *testing.T) {
 					"public_key":         "test-public-key",
 					"format":             "%h %l %u %t \"%r\" %>s %b",
 					"format_version":     uint(2),
-					"gzip_level":         uint(0),
+					"gzip_level":         uint8(0),
 					"message_type":       "classic",
 					"placement":          "waf_debug",
 					"response_condition": "error_response",
