+++ fastly/block_fastly_service_logging_cloudfiles_test.go
@@ -48,7 +48,7 @@ func TestResourceFastlyFlattenCloudfiles(t *testing.T) {
 					"public_key":         pgpPublicKey(t),
 					"format":             "%h %l %u %t \"%r\" %>s %b",
 					"format_version":     uint(2),
-					"gzip_level":         uint(0),
+					"gzip_level":         uint8(0),
 					"message_type":       "classic",
 					"path":               "/",
 					"region":             "ORD",
